Complete add book to stack functionality
