+++ models/stack.js
@@ -1,17 +1,18 @@
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 
-
 const stackSchema = new Schema(
 	{
 		name: {
 			type: String,
 			required: true,
 		},
-		books: {
-			type: Schema.Types.ObjectId, 
-			ref:"Book"
-		},
+		books: [
+			{
+				type: Schema.Types.ObjectId,
+				ref: "Book",
+			},
+		],
 	},
 	{ timestamps: true }
 );
